resampler-test: Add zero padding.
For the resampler test we create 10 samples and run them through a
resampler and then we try to run run the resampler backwards. This
crashes if our original 10 samples are not enough to produce any output
on the original run.

This patch adds additional zero padding to our original 10 samples. This
way all resample methods produce enough output to run the resampler
backwards.