resampler-test: Change the default rates
The default test case in resampler-test is to run the auto resample
method with some test samples from a rate of 44100Hz to a rate of
44100Hz.

When the rates are equal our resamlper automatically chooses to use the
'copy' resamlpe method (that is no resampling is done). This is quite
useless as our default test case (which is run by make check), so change
the rates to something more sensible.