resampler-test: Change the default test case to test all methods
This changes the default behaviour of our resampler-test if we
don't provide a resmaple method or sample format on the command line.

If neither a method or a sample format is specified we test all
resampling methods.

This way if we run 'make check' all our resampling methods will be
tested.